requirements.txt
fastapi
uvicorn[standard]
python-escpos
Pillow
pydantic-settings
python-dotenv
python-multipart
pyusb
qrcode
python-barcode